+++ zetaclient/chains/base/observer.go
@@ -33,6 +33,9 @@ const (
 
 	// MonitoringErrHandlerRoutineTimeout is the timeout for the handleMonitoring routine that waits for an error from the monitorVote channel
 	MonitoringErrHandlerRoutineTimeout = 5 * time.Minute
+  
+	// defaultSampledLogInterval is the default interval for sampled logs
+	defaultSampledLogInterval = 10
 )
 
 // Observer is the base structure for chain observers, grouping the common logic for each chain observer client.
@@ -428,5 +431,6 @@ func newObserverLogger(chain chains.Chain, logger Logger) ObserverLogger {
 		Inbound:    log.With().Str(logs.FieldModule, logs.ModNameInbound).Logger(),
 		Outbound:   log.With().Str(logs.FieldModule, logs.ModNameOutbound).Logger(),
 		Compliance: complianceLog,
+		Sampled:    log.Sample(&zerolog.BasicSampler{N: defaultSampledLogInterval}),
 	}
 }