+++ zetaclient/chains/base/observer_test.go
@@ -55,6 +55,7 @@ func newTestSuite(t *testing.T, chain chains.Chain, opts ...opt) *testSuite {
 
 	// constructor parameters
 	chainParams := *sample.ChainParams(chain.ChainId)
+	chainParams.IsSupported = true
 	chainParams.ConfirmationParams = &observertypes.ConfirmationParams{
 		SafeInboundCount:  defaultConfirmationCount,
 		SafeOutboundCount: defaultConfirmationCount,
@@ -210,6 +211,7 @@ func TestObserverGetterAndSetter(t *testing.T) {
 		logger.Inbound.Info().Msg("print inbound log")
 		logger.Outbound.Info().Msg("print outbound log")
 		logger.Compliance.Info().Msg("print compliance log")
+		logger.Sampled.Info().Msg("print sampled log")
 	})
 }