+++ zetaclient/config/config_chain.go
@@ -26,6 +26,7 @@ func New(setDefaults bool) Config {
 	}
 
 	if setDefaults {
+		cfg.MempoolCongestionThreshold = DefaultMempoolCongestionThreshold
 		cfg.EVMChainConfigs = evmChainsConfigs()
 		cfg.BTCChainConfigs = btcChainsConfigs()
 		cfg.SolanaConfig = solanaConfigLocalnet()