+++ zetaclient/zetacore/tx_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 
 	"cosmossdk.io/math"
+	abcitypes "github.com/cometbft/cometbft/abci/types"
+	coretypes "github.com/cometbft/cometbft/rpc/core/types"
 	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
@@ -272,9 +274,10 @@ func TestZetacore_MonitorVoteInboundResult(t *testing.T) {
 	ctx := context.Background()
 
 	address := sdktypes.AccAddress(mocks.TestKeyringPair.PubKey().Address().Bytes())
+	cometbftMock := mocks.NewSDKClientWithErr(t, nil, 0)
 	client := setupZetacoreClient(t,
 		withObserverKeys(keys.NewKeysWithKeybase(mocks.NewKeyring(), address, testSigner, "")),
-		withCometBFT(mocks.NewSDKClientWithErr(t, nil, 0)),
+		withCometBFT(cometbftMock),
 		withAccountRetriever(t, 5, 4),
 	)
 
@@ -284,6 +287,63 @@ func TestZetacore_MonitorVoteInboundResult(t *testing.T) {
 		}, nil)
 
 		require.NoError(t, err)
+	})
+
+	t.Run("monitor inbound vote out of gas", func(t *testing.T) {
+		// ARRANGE
+		msg := &crosschaintypes.MsgVoteInbound{}
+
+		// given mock tx result
+		cometbftMock.SetTxResult(&coretypes.ResultTx{
+			TxResult: abcitypes.ExecTxResult{
+				Log:       "out of gas",
+				GasWanted: 1000,
+			},
+		})
+
+		// should not adjust the gas limit, no out-of-gas failure so far
+		gasLimit, retryGasLimit, retryable := client.getAdjustedGasLimitForInboundVote(msg.Digest(), 1000, 2000)
+		require.Equal(t, uint64(1000), gasLimit)
+		require.Equal(t, uint64(2000), retryGasLimit)
+		require.True(t, retryable)
+
+		// ACT 1
+		// monitor out of gas error, should record it
+		err := client.MonitorVoteInboundResult(ctx, sampleHash, 0, msg)
+		require.NoError(t, err)
+
+		// it is okay to monitor same tx again, nothing will happen
+		err = client.MonitorVoteInboundResult(ctx, sampleHash, 0, msg)
+		require.NoError(t, err)
+
+		// ASSERT 1.1
+		// should adjust the gas limit
+		gasLimit, retryGasLimit, retryable = client.getAdjustedGasLimitForInboundVote(msg.Digest(), 1000, 2000)
+		require.Equal(t, uint64(2000), gasLimit)
+		require.Zero(t, retryGasLimit)
+		require.True(t, retryable)
+
+		// ASSERT 1.2
+		// not retryable, because given gas limits are too low
+		gasLimit, retryGasLimit, retryable = client.getAdjustedGasLimitForInboundVote(msg.Digest(), 500, 1000)
+		require.Zero(t, gasLimit)
+		require.Zero(t, retryGasLimit)
+		require.False(t, retryable)
+
+		// set mock tx result to nil, so tx is successful
+		cometbftMock.SetTxResult(nil)
+
+		// ACT 2
+		// monitor tx as a successful, should remove the record of out of gas failure
+		err = client.MonitorVoteInboundResult(ctx, sampleHash, 0, msg)
+		require.NoError(t, err)
+
+		// ASSERT 2.1
+		// should not adjust the gas limit
+		gasLimit, retryGasLimit, retryable = client.getAdjustedGasLimitForInboundVote(msg.Digest(), 1000, 2000)
+		require.Equal(t, uint64(1000), gasLimit)
+		require.Equal(t, uint64(2000), retryGasLimit)
+		require.True(t, retryable)
 	})
 }